Addition of 2 Derived Quantities
I have added the ability to copy the 'analysis' group to the auxiliary/reduced h5 file. In addition, I have added two derived quantities, a 2-D shock contour and average atomic mass in amu. The shock contour data is not written into the xmf/xml, but rather it is added to the secondary h5 file directly for use by the pyplotter to overlay a contour of detected shocks. The average atomic mass (denoted Abar) is added to the xmf/xml tree creating a new plot-able variable for the pyplotter.